3.04 twit Update

+++ src/components/Twit.js
@@ -1,21 +1,44 @@
 import { dbService } from "fbase";
-import React from "react";
+import React,{useState} from "react";
 
 const Twit=({twitObj, isOwner})=>{
-    const onDelete=()=>{
+    const [twit,setTwit]=useState(twitObj.text);
+    const [toggle,setToggle]=useState(false);
+    const onDelete=async ()=>{
         const conFirm=window.confirm("You want to Delete??");
-        if(conFirm)
-            dbService.doc(`/twit/${twitObj.id}`).delete();
+        conFirm && await dbService.doc(`/twit/${twitObj.id}`).delete();
+    }
+    const onUpdate=()=>setToggle(prev=>!prev);
+    const onChange=e=>setTwit(e.target.value);
+    const onCancel=e=> setToggle(!toggle);
+    const onSubmit=async e=>{
+        e.preventDefault();
+        await dbService.doc(`/twit/${twitObj.id}`).update({
+            text:twit
+        });
+        setToggle(prev=>!prev);
     }
     return(
     <div key={twitObj.id}>
-        <h4>{twitObj.text}</h4>
-        {isOwner &&
+    {toggle ? 
         <>
-        <button>Edit</button>
-        <button onClick={onDelete}>Delete</button>
+        <form onSubmit={onSubmit}>
+            <input type="text" placeholder={twit} onChange={onChange} value={twit}/>
+            <input type="submit" value="ReTwit"/>
+            <button onClick={onCancel}>Cancel</button>
+        </form>
         </>
+        :
+        <>
+            <h4>{twitObj.text}</h4>
+        {isOwner &&
+            <>
+            <button onClick={onUpdate}>Edit</button>
+            <button onClick={onDelete}>Delete</button>
+            </>
         }
+        </>
+    }
     </div>
     )
 }
